+++ txbits/app/securesocial/core/Authorization.scala
@@ -39,7 +39,7 @@ trait Authorization {
 }
 
 object Authorization2fa extends Authorization {
-  def verifyTokenForUser(user: BigDecimal, token: String, secret: String) = {
+  def verifyTokenForUser(user: Long, token: String, secret: String) = {
     if (TOTPAuthenticator.pinMatchesSecret(token, TOTPSecret(secret))) {
       if (!userModel.TOTPTokenIsBlacklisted(user, token)) {
         val blacklistDuration: Long = Play.current.configuration.getLong("TOTPTokenBlacklistDurationMs").getOrElse(30 * 7 * 1000)
